requirements.txt
python-nmap>=0.7.1
scapy>=2.4.0
psutil>=5.8.0
flask>=2.0.0
requests>=2.25.0
netifaces>=0.10.0
pymongo>=4.0.0
motor>=3.0.0
dnspython>=2.0.0
python-dotenv>=0.19.0
